test(stella-cli): split model_catalog.rs's inline tests into a sibling file
model_catalog.rs sat 3 lines from the strict 1500-line god-file ceiling,
with ~318 of its lines being an inline #[cfg(test)] mod tests block. Move
the block verbatim into model_catalog/tests.rs, following the same
sibling-tests.rs pattern already used 18 times in this crate. Pure test
move, no production change.

Closes #3700
